backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
click==8.3.0
fastapi==0.118.0
h11==0.16.0
idna==3.10
pydantic==2.11.10
pydantic-settings==2.7.1
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
SQLAlchemy==2.0.36
psycopg[binary]==3.2.10
passlib[bcrypt]==1.7.4
python-jose==3.3.0
